Add Prefixing, Postfixing to Query Output
Sigma originally had no plans to add literal query prefixes or postfixes
to generated query outputs, so it had to be added. Reason for including them
is to better control whitelisting, at least on a per-backend basis.

This was achieved by adding two new fields to the Sigma rule's YAML
Specification: search_prefix and search_postfix.

This, of course, comes at a cost of portability to other search tools,
but for our organization, we decided that there's always going to be a
mountain of technical debt to repay when switching platforms, and the
benefits of including whitelisting within this  platform far outweighs
the technical debt incurred.